refactor: use all validation code
This adds some missing validation steps by using a single validation
entrypoint.
